Cleaning up demo
